Added caveat for role renaming
It appears that renaming the default roles to the expected LDAP group matches does not have the intended outcome.  Creating new roles works, though. For example, renaming the default BookStack role "Editor" to "users" should map to the LDAP group "users". But it does not. Creating a new BookStack role "users", however, works as intended.